feat: 3차 코드 반영
