+++ apps/client/src/pages/my/hooks/use-performance-detail.ts
@@ -1,5 +1,5 @@
 import { useSuspenseQuery } from '@tanstack/react-query';
-import { PERFORMANCE_QUERY_OPTIONS } from '@shared/apis/my-confeti';
+import { PERFORMANCE_QUERY_OPTIONS } from '@shared/apis/user-queries';
 
 export const useUserPerformanceMore = () => {
   const { data } = useSuspenseQuery(PERFORMANCE_QUERY_OPTIONS.ALL());
