+++ apps/client/src/pages/my/page/profile/my-profile.tsx
@@ -9,7 +9,7 @@ import NoConfetiSection from '@pages/my/components/confeti/no-confeti-section';
 import ArtistSection from '@pages/my/components/artist/artist-section';
 import UserInfo from '@pages/my/components/profile/user-info';
 import ConfetiSection from '@pages/my/components/confeti/confeti-section';
-import { PERFORMANCE_QUERY_OPTIONS } from '@shared/apis/my-confeti';
+import { PERFORMANCE_QUERY_OPTIONS } from '@shared/apis/user-queries';
 import { useSuspenseQuery } from '@tanstack/react-query';
 
 const MyProfile = () => {
